+++ src/packages/core/menu/components/menu-item/index.ts
@@ -1,1 +1,2 @@
 export * from './menu-item-default.element.js';
+export * from './link/index.js';
