+++ src/packages/core/menu/components/menu-item/link/index.ts
@@ -1,0 +1,1 @@
+export * from './types.js';
