+++ src/packages/core/menu/components/menu-item/link/types.ts
@@ -1,0 +1,17 @@
+import type { ManifestMenuItem, MetaMenuItem } from '../../../menu-item.extension.js';
+
+export interface ManifestMenuItemLinkKind extends ManifestMenuItem {
+	type: 'menuItem';
+	kind: 'link';
+	meta: MetaMenuItemLinkKind;
+}
+
+export interface MetaMenuItemLinkKind extends MetaMenuItem {
+	href: string;
+}
+
+declare global {
+	interface UmbExtensionManifestMap {
+		umbLinkMenuItemKind: ManifestMenuItemLinkKind;
+	}
+}
