+++ src/packages/core/tree/tree-menu-item-default/index.ts
@@ -1,1 +1,2 @@
 export * from './tree-menu-item-default.element.js';
+export * from './types.js';
